Move platform-agnostic logic to its own file
Move the platform-agnostic code to its own file, so that someone
running the slackbot doesn't need to have telegram libraries
installed (and vice versa).  Left the telegram run code in the
same file, for backwards compatibility.
